requirements.txt
pillow==11.1.0
rarfile==4.2
